+++ packages/convergence/CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.5.0] - 2018-03-14
+
 ### Added
 
 - full support for returning promises in `.do()` blocks
 - timeout error thrown when convergence exceeds timeout
 - support for returning convergence in `.do()` blocks
+- `isConvergence` helper function
 
 ## [0.4.0] - 2018-03-02